[CSM-423] Move Test.Util to cardano-sl
