+++ lib/cardano-sl.cabal
@@ -135,6 +135,7 @@ library
 
                         -- Common code for tests
                         Test.Pos.Configuration
+                        Test.Pos.Util
                         Test.Pos.Block.Logic.Emulation
                         Test.Pos.Block.Logic.Mode
                         Test.Pos.Block.Logic.Util
@@ -483,7 +484,7 @@ test-suite cardano-test
                        Test.Pos.Update.Identity.SafeCopySpec
                        Test.Pos.Update.MemStateSpec
                        Test.Pos.Update.PollSpec
-                       Test.Pos.Util
+                       Test.Pos.Helpers
                        Test.Pos.Util.LimitsSpec
                        Test.Pos.Util.ModifierSpec
                        Test.Pos.UtilSpec
