+++ lib/test/Test/Pos/Block/Identity/BinarySpec.hs
@@ -13,7 +13,8 @@ import qualified Pos.Block.Core      as BT
 import qualified Pos.Block.Network   as BT
 import qualified Pos.Communication   ()
 
-import           Test.Pos.Util       (binaryTest, withDefConfiguration)
+import           Test.Pos.Helpers    (binaryTest)
+import           Test.Pos.Util       (withDefConfiguration)
 
 spec :: Spec
 spec = withDefConfiguration $ describe "Block types" $ do
