+++ lib/test/Test/Pos/Communication/Identity/BinarySpec.hs
@@ -14,7 +14,7 @@ import           Pos.Arbitrary.Infra   ()
 import qualified Pos.Communication     as C
 
 import           Test.Pos.CborSpec     (extensionProperty)
-import           Test.Pos.Util         (binaryTest)
+import           Test.Pos.Helpers      (binaryTest)
 
 spec :: Spec
 spec = describe "Communication" $ do
