+++ lib/test/Test/Pos/Helpers.hs
@@ -1,18 +1,10 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE RankNTypes          #-}
 
-module Test.Pos.Util
-       ( HasStaticConfigurations
-       , withDefConfiguration
-       , withDefInfraConfiguration
-       , withDefNodeConfiguration
-       , withDefGtConfiguration
-       , withDefUpdateConfiguration
-       , withDefConfigurations
-       , withStaticConfigurations
-
+module Test.Pos.Helpers
+       (
        -- * From/to
-       , binaryEncodeDecode
+         binaryEncodeDecode
        , binaryTest
        , safeCopyEncodeDecode
        , safeCopyTest
@@ -21,6 +13,8 @@ module Test.Pos.Util
        , showReadId
        , showReadTest
        , canonicalJsonTest
+
+       , splitIntoChunks
 
        -- * Message length
        , msgLenLimitedTest
@@ -36,19 +30,7 @@ module Test.Pos.Util
        , formsMonoid
        , formsCommutativeMonoid
 
-       -- * Monadic properties
-       , stopProperty
-       , maybeStopProperty
-       , splitIntoChunks
-
        -- * Various properties and predicates
-       , qcIsJust
-       , qcIsNothing
-       , qcIsLeft
-       , qcIsRight
-       , qcElem
-       , qcNotElem
-       , qcFail
        , blockPropertySpec
        ) where
 
@@ -59,7 +41,6 @@ import qualified Data.ByteString                  as BS
 import           Data.SafeCopy                    (SafeCopy, safeGet, safePut)
 import qualified Data.Semigroup                   as Semigroup
 import           Data.Serialize                   (runGet, runPut)
-import           Data.Tagged                      (Tagged (..))
 import           Data.Typeable                    (typeRep)
 import           Formatting                       (formatToString, int, (%))
 import           Prelude                          (read)
@@ -71,65 +52,16 @@ import           Test.QuickCheck                  (Arbitrary (arbitrary), Proper
                                                    property, resize, suchThat, vectorOf,
                                                    (.&&.), (===))
 import           Test.QuickCheck.Gen              (choose)
-import           Test.QuickCheck.Monadic          (PropertyM, pick, stop)
-import           Test.QuickCheck.Property         (Result (..), failed)
+import           Test.QuickCheck.Monadic          (PropertyM, pick)
 import qualified Text.JSON.Canonical              as CanonicalJSON
 
 import           Pos.Binary                       (AsBinaryClass (..), Bi (..), serialize,
                                                    serialize', unsafeDeserialize)
 import           Pos.Communication                (Limit (..), MessageLimitedPure (..))
-import           Pos.Configuration                (HasNodeConfiguration,
-                                                   withNodeConfiguration)
-import           Pos.Core                         (HasConfiguration, withGenesisSpec)
-import           Pos.Infra.Configuration          (HasInfraConfiguration,
-                                                   withInfraConfiguration)
-import           Pos.Ssc.GodTossing.Configuration (HasGtConfiguration,
-                                                   withGtConfiguration)
-import           Pos.Update.Configuration         (HasUpdateConfiguration,
-                                                   withUpdateConfiguration)
-
-import           Pos.Launcher.Configuration       (Configuration (..), HasConfigurations)
-
+import           Pos.Configuration                (HasNodeConfiguration)
+import           Pos.Core                         (HasConfiguration)
+import           Pos.Ssc.GodTossing.Configuration (HasGtConfiguration)
 import           Test.Pos.Block.Logic.Mode        (BlockProperty, blockPropertyTestable)
-import           Test.Pos.Configuration           (defaultTestConf)
-
--- | This constraint requires all configurations which are not
--- always hardcoded in tests (currently).
-type HasStaticConfigurations =
-    ( HasInfraConfiguration
-    , HasUpdateConfiguration
-    , HasGtConfiguration
-    , HasNodeConfiguration
-    )
-
-withDefNodeConfiguration :: (HasNodeConfiguration => r) -> r
-withDefNodeConfiguration = withNodeConfiguration (ccNode defaultTestConf)
-
-withDefGtConfiguration :: (HasGtConfiguration => r) -> r
-withDefGtConfiguration = withGtConfiguration (ccGt defaultTestConf)
-
-withDefUpdateConfiguration :: (HasUpdateConfiguration => r) -> r
-withDefUpdateConfiguration = withUpdateConfiguration (ccUpdate defaultTestConf)
-
-withDefInfraConfiguration :: (HasInfraConfiguration => r) -> r
-withDefInfraConfiguration = withInfraConfiguration (ccInfra defaultTestConf)
-
-withDefConfiguration :: (HasConfiguration => r) -> r
-withDefConfiguration = withGenesisSpec 0 (ccCore defaultTestConf)
-
-withStaticConfigurations :: (HasStaticConfigurations => r) -> r
-withStaticConfigurations patak =
-    withDefNodeConfiguration $
-    withDefGtConfiguration $
-    withDefUpdateConfiguration $
-    withDefInfraConfiguration patak
-
-withDefConfigurations :: (HasConfigurations => r) -> r
-withDefConfigurations bardaq =
-    withDefConfiguration $ withStaticConfigurations bardaq
-
-instance Arbitrary a => Arbitrary (Tagged s a) where
-    arbitrary = Tagged <$> arbitrary
 
 ----------------------------------------------------------------------------
 -- From/to tests
@@ -279,7 +211,6 @@ msgLenLimitedTest
     => Spec
 msgLenLimitedTest = msgLenLimitedTest' @a msgLenLimit "" (const True)
 
-
 ----------------------------------------------------------------------------
 -- Monoid/Semigroup laws
 ----------------------------------------------------------------------------
@@ -344,26 +275,6 @@ shouldThrowException
 shouldThrowException action exception arg =
     (return $! action arg) `shouldThrow` exception
 
-----------------------------------------------------------------------------
--- Monadic testing
-----------------------------------------------------------------------------
-
--- Note, 'fail' does the same thing, but:
--- • it's quite trivial, almost no copy-paste;
--- • it's 'fail' from 'Monad', not 'MonadFail';
--- • I am not a fan of 'fail'.
--- | Stop 'PropertyM' execution with given reason. The property will fail.
-stopProperty :: Monad m => Text -> PropertyM m a
-stopProperty msg = stop failed {reason = toString msg}
-
--- | Use 'stopProperty' if the value is 'Nothing' or return something
--- it the value is 'Just'.
-maybeStopProperty :: Monad m => Text -> Maybe a -> PropertyM m a
-maybeStopProperty msg =
-    \case
-        Nothing -> stopProperty msg
-        Just x -> pure x
-
 -- | Split given list into chunks with size up to given value.
 splitIntoChunks :: Monad m => Word -> [a] -> PropertyM m [NonEmpty a]
 splitIntoChunks 0 _ = error "splitIntoChunks: maxSize is 0"
@@ -377,40 +288,6 @@ splitIntoChunks maxSize items = do
 ----------------------------------------------------------------------------
 -- Various properties and predicates
 ----------------------------------------------------------------------------
-
-qcIsJust :: Maybe a -> Property
-qcIsJust (Just _) = property True
-qcIsJust Nothing  = qcFail "expected Just, got Nothing"
-
-qcIsNothing :: Show a => Maybe a -> Property
-qcIsNothing Nothing  = property True
-qcIsNothing (Just x) = qcFail ("expected Nothing, got Just (" <> show x <> ")")
-
-qcIsLeft :: Show b => Either a b -> Property
-qcIsLeft (Left _)  = property True
-qcIsLeft (Right x) = qcFail ("expected Left, got Right (" <> show x <> ")")
-
-qcIsRight :: Show a => Either a b -> Property
-qcIsRight (Right _) = property True
-qcIsRight (Left x)  = qcFail ("expected Right, got Left (" <> show x <> ")")
-
-qcElem
-    :: (Eq a, Show a, Show t, NontrivialContainer t, Element t ~ a)
-    => a -> t -> Property
-qcElem x xs =
-    counterexample ("expected " <> show x <> " to be in " <> show xs) $
-    x `elem` xs
-
-qcNotElem
-    :: (Eq a, Show a, Show t, NontrivialContainer t, Element t ~ a)
-    => a -> t -> Property
-qcNotElem x xs =
-    counterexample ("expected " <> show x <> " not to be in " <> show xs) $
-    not (x `elem` xs)
-
--- | A property that is always false
-qcFail :: Text -> Property
-qcFail s = counterexample (toString s) False
 
 -- | Specialized version of 'prop' function from 'hspec'.
 blockPropertySpec ::
